Moved signUpUser function to authService, deleted userService
