+++ src/services/authService.js
@@ -22,3 +22,20 @@ export const loginUser = async (user) => {
     }
   }
 };
+
+export const signUpUser = async (user) => {
+  try {
+    const response = await axios.post(
+      'http://localhost:3003/api/v1/usuarios',
+      user,
+    );
+    return response.data;
+  } catch (error) {
+    if (error.response) {
+      const errorResponse = error.response.data.mensagem.texto;
+      return { status: false, mensagem: errorResponse };
+    } else {
+      return { status: false, mensagem: { texto: error.message } };
+    }
+  }
+};
